Add support for BLE battery service
This adds a battery level that complies with the BLE Battery Service
specification and can be monitored remotely by any BLE device.

Some esp32 boards use an ADC pin to read battery level while others
have an I2C connected battery charger chip, so your sketch is still
responsible for reading the battery level and providing it to the
gadget service.

The battery level must be a percentage from 0-100, not a voltage.

    uint8_t battery_percent = get_battery_percent_from_your_board();
    gadgetBle.writeBatteryLevel(battery_percent);

The battery service shows up in nRF Connect and Light Blue, however
the MyAmbience app does not currently show the battery level, likely
because it doesn't yet expect the esp32 gadgets to have a battery
service.

Notifications are sent when the battery level actually changes as long
as a download from MyAmbience is not currently in progress.

Ref: #5