OpenCO2 mini
Store wifi credentials via preferences
Add the option to load the wifi credentials on boot. Before this change the user needed to input the wifi ssid/password every time the device was rebooted. Tested on a ESP32-S3.

pNimBLEAdvertising interval of 1s
getConneted() to check if client communication is active